api/pyproject.toml
[tool.poetry]
name = "paper-recommender"
version = "0.1.0"
description = ""
authors = ["Ryusei Kakujo <ryusei.kakujo@traffine.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.1"
pydantic = "^1.10.5"
tenacity = "^8.2.1"
numpy = "^1.24.2"
python-multipart = "^0.0.6"
arrow = "^1.2.3"
setuptools = "^68.0.0"
tiktoken = "^0.4.0"
filelock = "^3.12.2"
boto3 = "^1.28.65"
mecab-python3 = "^1.0.8"
ipadic = "^1.0.0"
openai = "^1.10.0"
loguru = "^0.7.2"
pinecone-client = "^3.2.2"

[tool.poetry.scripts]
start = "server.main:start"
dev = "local-server.main:start"

[tool.poetry.group.dev.dependencies]
httpx = "^0.23.3"
streamlit = "^1.33.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode="auto"
